Phase 1 Complete: Repository setup and basic CLI structure
- Renamed ct/ to zt/ and updated module to github.com/cpepper96/zarf-testing
- Added Zarf dependency and removed Helm dependency
- Updated CLI commands to use 'zt' with Zarf-specific descriptions
- Updated flags from chart-dirs to zarf-dirs and chart references to package references
- Implemented command stubs with 'not implemented' messages for lint and list-changed
- All commands compile and show appropriate help text
- Version command working

Tasks completed:
- Task 1.1: Repository Setup & Initial Adaptation
- Task 1.2: Basic Command Structure

Next: Task 2.1 (Package Discovery) and Task 2.2 (Zarf SDK Integration)
Amp-Thread-ID: https://ampcode.com/threads/T-1e7b51ee-c7d5-4a09-91dc-5022cb172f80
