--- ct/cmd/docGen.go
@@ -1,47 +1,0 @@
-// Copyright The Helm Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/MakeNowJust/heredoc"
-	"github.com/spf13/cobra"
-	"github.com/spf13/cobra/doc"
-)
-
-func newGenerateDocsCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "doc-gen",
-		Short: "Generate documentation",
-		Long: heredoc.Doc(`
-			Generate documentation for all commands
-			to the 'docs' directory.`),
-		Hidden: true,
-		RunE:   generateDocs,
-	}
-}
-
-func generateDocs(_ *cobra.Command, _ []string) error {
-	fmt.Println("Generating docs...")
-
-	err := doc.GenMarkdownTree(NewRootCmd(), "doc")
-	if err != nil {
-		return err
-	}
-
-	fmt.Println("Done.")
-	return nil
-}
