+++ pkg/chart/integration_test.go
@@ -23,10 +23,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/helm/chart-testing/v3/pkg/config"
-	"github.com/helm/chart-testing/v3/pkg/exec"
-	"github.com/helm/chart-testing/v3/pkg/tool"
-	"github.com/helm/chart-testing/v3/pkg/util"
+	"github.com/cpepper96/zarf-testing/pkg/config"
+	"github.com/cpepper96/zarf-testing/pkg/exec"
+	"github.com/cpepper96/zarf-testing/pkg/tool"
+	"github.com/cpepper96/zarf-testing/pkg/util"
 	"github.com/stretchr/testify/assert"
 )
