+++ pkg/config/config.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/mitchellh/go-homedir"
 
-	"github.com/helm/chart-testing/v3/pkg/util"
+	"github.com/cpepper96/zarf-testing/pkg/util"
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 	"github.com/spf13/viper"
