+++ pkg/exec/exec.go
@@ -22,7 +22,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/helm/chart-testing/v3/pkg/util"
+	"github.com/cpepper96/zarf-testing/pkg/util"
 )
 
 type ProcessExecutor struct {
