+++ pkg/tool/git.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/helm/chart-testing/v3/pkg/exec"
+	"github.com/cpepper96/zarf-testing/pkg/exec"
 )
 
 type Git struct {
