+++ pkg/tool/kubectl.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-retryablehttp"
-	"github.com/helm/chart-testing/v3/pkg/exec"
+	"github.com/cpepper96/zarf-testing/pkg/exec"
 )
 
 type Kubectl struct {
